Refactor all the splitted method and added comments to it
