+++ elements/layercontrol/src/methods/add-layer/index.js
@@ -1,6 +1,8 @@
-export { default as jsonInputChangeMethod } from "./json-input-change";
-export { default as addLayerMethod } from "./add-layer";
-export { default as urlInputChangeMethod } from "./url-input-change";
-export { default as wmsSearchURLMethod } from "./wms-search-url";
-export { default as urlLayerMethod } from "./url-layer";
-export { default as openCloseTabMethod } from "./open-close-tab";
+// Importing methods from respective files and exporting them
+
+export { default as jsonInputChangeMethod } from "./json-input-change"; // Handles changes in the input field, parsing entered data into JSON format.
+export { default as addLayerMethod } from "./add-layer"; // Handles the addition of one or multiple layers
+export { default as urlInputChangeMethod } from "./url-input-change"; // Handles the change event in the URL input field
+export { default as wmsSearchURLMethod } from "./wms-search-url"; // Handles the URL search process
+export { default as urlLayerMethod } from "./url-layer"; // Handles changes in the URL input field
+export { default as openCloseTabMethod } from "./open-close-tab"; // Handles the opening or closing of tabs
