+++ elements/layercontrol/src/methods/layer-config/index.js
@@ -1,1 +1,3 @@
-export { default as dataChangeMethod } from "./data-change";
+// Importing methods from respective files and exporting them
+
+export { default as dataChangeMethod } from "./data-change"; // Handles data change with help of new tile url function
