+++ elements/layercontrol/src/methods/layer-list/index.js
@@ -1,2 +1,4 @@
-export { default as firstUpdatedMethod } from "./first-updated";
-export { default as updateMethod } from "./update";
+// Importing methods from respective files and exporting them
+
+export { default as firstUpdatedMethod } from "./first-updated"; // Handles initial settings and event triggers upon the first update of a component.
+export { default as updateMethod } from "./update"; // Manages changes in the length of layers
