+++ elements/layercontrol/src/methods/layer-list/update.js
@@ -1,13 +1,16 @@
+// Importing the lodash debounce function for handling debouncing
 import _debounce from "lodash.debounce";
 
 /**
- * Handles initial settings and event triggers upon the component's first update.
+ * Handles updating and triggering events when the layers change length upon component update.
  *
- * @param {import("../../components/layerList").EOxLayerControlLayerList} EOxLayerControlLayerList
+ * @param {import("../../components/layerList").EOxLayerControlLayerList} EOxLayerControlLayerList - The EOxLayerControlLayerList object containing layers.
  */
 const updateMethod = (EOxLayerControlLayerList) => {
+  // Destructuring layers from the EOxLayerControlLayerList object
   const { layers } = EOxLayerControlLayerList;
 
+  // Debounced function for handling changes in layers' length
   const debHandleLayersChangeLength = _debounce(() => {
     EOxLayerControlLayerList.requestUpdate();
     EOxLayerControlLayerList.dispatchEvent(
@@ -15,13 +18,17 @@ const updateMethod = (EOxLayerControlLayerList) => {
     );
   }, 50);
 
+  // Function to handle changes in layers' length
   const handleLayersChangeLength = () => debHandleLayersChangeLength();
 
+  // If layers do not exist, exit the function
   if (!layers) return;
 
+  // Unregister previous listeners to prevent duplicates
   if (layers.hasListener("change:length"))
     layers?.un("change:length", handleLayersChangeLength);
 
+  // Register a listener for changes in the layers' length
   layers.on("change:length", handleLayersChangeLength);
 };
