+++ elements/layercontrol/src/methods/layer/index.js
@@ -1,2 +1,4 @@
-export { default as firstUpdatedMethod } from "./first-updated";
-export { default as inputClickMethod } from "./input-click";
+// Importing methods from respective files and exporting them
+
+export { default as firstUpdatedMethod } from "./first-updated"; // Check and update layer zoom visibility at beginning
+export { default as inputClickMethod } from "./input-click"; // Handles the click event on an input element to control layer visibility.
