+++ elements/layercontrol/src/methods/layercontrol/first-update.js
@@ -1,11 +1,12 @@
 /**
- * Handles initial settings and event triggers upon the component's first update.
+ * Retrieves the EOxMap element based on the provided selector in EOxLayerControl's 'for' property.
  *
- * @param {import("../../main").EOxLayerControl} EOxLayerControl
- * @return {import("@eox/map/main").EOxMap}
+ * @param {import("../../main").EOxLayerControl} EOxLayerControl - The EOxLayerControl element.
+ * @returns {import("@eox/map/main").EOxMap | null} The EOxMap element or null if not found.
  */
 const firstUpdatedMethod = (EOxLayerControl) => {
-  return document.querySelector(EOxLayerControl.for);
+  const mapSelector = EOxLayerControl.for;
+  return document.querySelector(mapSelector);
 };
 
 export default firstUpdatedMethod;
