+++ elements/layercontrol/src/methods/layercontrol/index.js
@@ -1,5 +1,5 @@
 // Importing methods from respective files and exporting them
 
-export { default as layerListChangeMethod } from "./layer-list-change";
-export { default as firstUpdatedMethod } from "./first-update";
-export { default as updateMethod } from "./update";
+export { default as layerListChangeMethod } from "./layer-list-change"; // Handles changes in the layer list within the EOxLayerControl component.
+export { default as firstUpdatedMethod } from "./first-update"; // Initial settings and event triggers upon the component's first update.
+export { default as updateMethod } from "./update"; // Method updates specific information within the EOxLayerControl component.
