+++ elements/layercontrol/src/methods/optional-list/index.js
@@ -1,3 +1,3 @@
 // Importing methods from respective files and exporting them
 
-export { default as addToListMethod } from "./add-to-list";
+export { default as addToListMethod } from "./add-to-list"; // Handle add layer to list
